# swagger/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
  AuthToken:
    description: JWT Token
    example:
      token: Iste odit culpa illo reiciendis quo.
    properties:
      token:
        description: JWT Token
        example: Dolores vero rerum.
        type: string
    required:
    - token
    title: 'Mediatype identifier: application/vnd.authtoken+json'
    type: object
  Version:
    description: The current running version
    example:
      build_time: Soluta ducimus explicabo et dolores.
      commit: Est quaerat eaque quaerat.
    properties:
      build_time:
        description: The date when build
        example: Quaerat dolorem voluptate non omnis.
        type: string
      commit:
        description: Commit SHA this build is based on
        example: Beatae accusantium quis magni.
        type: string
    required:
    - commit
    - build_time
    title: 'Mediatype identifier: application/vnd.version+json'
    type: object
host: almighty.io
info:
  description: The next big thing
  license:
    name: Apache License Version 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  title: 'ALMighty: One to rule them all'
  version: "1.0"
paths:
  /login/authorize:
    get:
      description: Authorize with the ALM
      operationId: login#authorize
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AuthToken'
        "401":
          description: Unauthorized
      schemes:
      - http
      summary: authorize
  /version:
    get:
      description: |-
        Show current running version

        ** Required security scopes**:
          * `system`
      operationId: version#show
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Version'
      schemes:
      - http
      summary: show
produces:
- application/json
responses:
  OK:
    description: OK
    schema:
      $ref: '#/definitions/Version'
  Unauthorized:
    description: Unauthorized
schemes:
- http
securityDefinitions:
  jwt:
    description: |-
      JWT Token Auth

      **Token URL**: http://almighty.io/api/login/authorize
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"